revlist: alternative impl for threaded listing
As known, revlist operation is expensive in time
due to the RTT of object store and/or the amount
of files that go-git has to parse.

The backing EncodedObjectStorer should be able to
handle concurrent requests to use this new
implementation. That is identified by implementing
ConcurrentStorer, returning maximum concurrency
the storage can handle. At this moment, built-in
filesystem storage is not eligible since the cache
gets written concurrently with no lock.

This is useful in case the backing storage is not
local, NAS filesystem or database over network for
example, as every object queries requires full
network round-trip time on these storage, it slows
down the entire process a ton if not parallized.